fix: исправить парсинг тел запросов флотов
