+++ apps/api/src/dto/fleets.dto.ts
@@ -1,7 +1,12 @@
 // Назначение файла: DTO для флотов
-// Основные модули: express-validator
+// Основные модули: express-validator, services/wialon, utils/wialonLocator
 import { body } from 'express-validator';
-import { parseLocatorLink } from '../services/wialon';
+import { DEFAULT_BASE_URL } from '../services/wialon';
+import { parseLocatorLink } from '../utils/wialonLocator';
+
+function ensureLocatorLink(value: string) {
+  parseLocatorLink(value, DEFAULT_BASE_URL);
+}
 
 export class CreateFleetDto {
   static rules() {
@@ -11,7 +16,7 @@ export class CreateFleetDto {
         .isString()
         .notEmpty()
         .custom((value) => {
-          parseLocatorLink(value);
+          ensureLocatorLink(value);
           return true;
         }),
     ];
@@ -27,7 +32,7 @@ export class UpdateFleetDto {
         .isString()
         .notEmpty()
         .custom((value) => {
-          parseLocatorLink(value);
+          ensureLocatorLink(value);
           return true;
         }),
     ];
